requirements.txt
openai 
anthropic 
python-dotenv
streamlit